[3] Optimizing security for Postman
