+++ src/main/kotlin/com/queckly/app/controller/SystemController.kt
@@ -1,13 +1,16 @@
 package com.queckly.app.controller
 
+import org.springframework.http.MediaType
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import reactor.core.publisher.Mono
 
 @RestController
-@RequestMapping("/system")
+@RequestMapping("/")
 class SystemController {
-    @GetMapping
-    fun sayHello(): Mono<String> = Mono.just("Hello, Authenticated!")
+    @GetMapping(produces = [MediaType.APPLICATION_JSON_VALUE])
+    fun sayHello(): Mono<Greetings> = Mono.just(Greetings("Hello, Authenticated!"))
 }
+
+data class Greetings(val message: String)
